+++ dilogin/src/main/java/di/dilogin/discord/command/DiscordRegisterBungeeCommand.java
@@ -47,6 +47,9 @@ public class DiscordRegisterBungeeCommand implements DiscordCommand {
 	@SuppressWarnings("deprecation")
 	@Override
 	public void execute(String message, MessageReceivedEvent event) {
+		
+		if (!MainController.getDILoginController().isRegisterByDiscordCommandEnabled())
+			return;
 
 		event.getMessage().delete().delay(Duration.ofSeconds(20)).queue();
 		if (userDao.containsDiscordId(event.getAuthor().getIdLong())) {
